Helpers moved to the abs folder
Because the new version of Pure Data needs to have the abstrations and
helpers in the same folder to properly links everything together I moved
all the helpers the `abs` folder. I will use later on the `examples`
folder to store proper patches that demonstrate few possibilities that
can be done with this library.

Resolves: #10

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>